# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Here QA can test and frontend or mobile developers can get information of API endpoints.",
        "title": "Monolithic project API Endpoints",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/pipeline": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Here pipeline can be updated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline"
                ],
                "summary": "Update pipeline",
                "parameters": [
                    {
                        "description": "post info",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PipelineUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PipelineApiResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.DefaultResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Here pipeline can be created.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline"
                ],
                "summary": "Create pipeline",
                "parameters": [
                    {
                        "description": "post info",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PipelineCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PipelineApiResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.DefaultResponse"
                        }
                    }
                }
            }
        },
        "/pipeline/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Here all pipelines can be got.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline"
                ],
                "summary": "Get pipelines list",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "order_by_created_at",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "project_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "-1 0 1",
                        "name": "webhook_status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PipelineApiFindResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.DefaultResponse"
                        }
                    }
                }
            }
        },
        "/pipeline/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Here pipeline can be got.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline"
                ],
                "summary": "Get pipeline by key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PipelineApiResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.DefaultResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Here pipeline can be deleted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pipeline"
                ],
                "summary": "Delete pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DefaultResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.DefaultResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AudioTrack": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_url": {
                    "type": "string"
                },
                "lang_code": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "size_kb": {
                    "type": "number"
                }
            }
        },
...
        "models.DefaultResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                }
            }
        },
...
        "models.PipelineApiFindResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/models.PipelinesFindResponse"
                },
                "error_code": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                }
            }
        },
        "models.PipelineApiResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/models.PipelineResponse"
                },
                "error_code": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                }
            }
        },
        "models.PipelineCreateReq": {
            "type": "object",
            "properties": {
                "audio_tracks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AudioTrack"
                    }
                },
                "fail_description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_url": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                },
                "max_resolution": {
                    "type": "string"
                },
                "output_key": {
                    "type": "string"
                },
                "output_path": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "resolutions_string": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "size_kb": {
                    "type": "number"
                },
                "stage": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Subtitle"
                    }
                },
                "transcode_duration_secs": {
                    "type": "number"
                },
                "upload_duration_secs": {
                    "type": "number"
                }
            }
        },
        "models.PipelineResponse": {
            "type": "object",
            "properties": {
                "audio_tracks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AudioTrack"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "fail_description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_url": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                },
                "max_resolution": {
                    "type": "string"
                },
                "output_key": {
                    "type": "string"
                },
                "output_path": {
                    "type": "string"
                },
                "preparation_duration_seconds": {
                    "type": "number"
                },
                "project_id": {
                    "type": "string"
                },
                "resolutions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Resolution"
                    }
                },
                "resolutions_string": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "size_kb": {
                    "type": "number"
                },
                "stage": {
                    "type": "string"
                },
                "stage_status": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Subtitle"
                    }
                },
                "transcode_duration_secs": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "upload_duration_secs": {
                    "type": "number"
                },
                "video_duration": {
                    "type": "number"
                },
                "webhook_last_retry": {
                    "type": "string"
                },
                "webhook_retry_count": {
                    "type": "integer"
                },
                "webhook_status": {
                    "type": "boolean"
                }
            }
        },
        "models.PipelineUpdateReq": {
            "type": "object",
            "properties": {
                "fail_description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_url": {
                    "type": "string"
                },
                "max_resolution": {
                    "type": "string"
                },
                "output_key": {
                    "type": "string"
                },
                "output_path": {
                    "type": "string"
                },
                "preparation_duration_seconds": {
                    "type": "number"
                },
                "resolutions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Resolution"
                    }
                },
                "resolutions_string": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "size_kb": {
                    "type": "number"
                },
                "stage": {
                    "type": "string"
                },
                "stage_status": {
                    "type": "string"
                },
                "transcode_duration_secs": {
                    "type": "number"
                },
                "upload_duration_secs": {
                    "type": "number"
                },
                "video_duration": {
                    "type": "number"
                },
                "webhook_last_retry": {
                    "type": "string"
                },
                "webhook_retry_count": {
                    "type": "integer"
                },
                "webhook_status": {
                    "type": "boolean"
                }
            }
        },
        "models.PipelinesFindResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "pipelines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PipelineResponse"
                    }
                }
            }
        },
...
        "models.Resolution": {
            "type": "object",
            "properties": {
                "bitrate": {
                    "type": "string"
                },
                "measure": {
                    "type": "string"
                },
                "resolution": {
                    "type": "string"
                }
            }
        },
...
        "models.Subtitle": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_url": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "size_kb": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}